# apps/frontend/src/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/asignacion-transporte/vehiculos": {
      "post": {
        "description": "Crea un nuevo vehículo en el sistema con toda la información requerida",
        "operationId": "AsignacionTransporteController_createVehiculo",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos del vehículo a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehiculoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehículo creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VehiculoResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o vehículo ya existe"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Crear un nuevo vehículo",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "get": {
        "description": "Obtiene una lista paginada de todos los vehículos registrados",
        "operationId": "AsignacionTransporteController_findAllVehiculos",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página (por defecto: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página (por defecto: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehículos obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VehiculoListResponseDto"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener todos los vehículos",
        "tags": [
          "Asignación de Transporte"
        ]
      }
    },
    "/api/asignacion-transporte/vehiculos/{id}": {
      "get": {
        "description": "Obtiene la información detallada de un vehículo específico",
        "operationId": "AsignacionTransporteController_findVehiculoById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del vehículo",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehículo encontrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VehiculoResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vehículo no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener un vehículo por ID",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "patch": {
        "description": "Actualiza la información de un vehículo existente",
        "operationId": "AsignacionTransporteController_updateVehiculo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del vehículo",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Datos del vehículo a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehiculoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehículo actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VehiculoResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o vehículo ya existe"
          },
          "404": {
            "description": "Vehículo no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Actualizar un vehículo",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "delete": {
        "description": "Elimina un vehículo del sistema (cambia su estado a inactivo)",
        "operationId": "AsignacionTransporteController_removeVehiculo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del vehículo",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehículo eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EmptyResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Vehículo no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Eliminar un vehículo",
        "tags": [
          "Asignación de Transporte"
        ]
      }
    },
    "/api/asignacion-transporte/conductores": {
      "post": {
        "description": "Crea un nuevo conductor en el sistema con toda la información requerida",
        "operationId": "AsignacionTransporteController_createConductor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos del conductor a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConductorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conductor creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConductorResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o conductor ya existe"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Crear un nuevo conductor",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "get": {
        "description": "Obtiene una lista paginada de todos los conductores registrados",
        "operationId": "AsignacionTransporteController_findAllConductores",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página (por defecto: 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Elementos por página (por defecto: 10)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conductores obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConductorListResponseDto"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener todos los conductores",
        "tags": [
          "Asignación de Transporte"
        ]
      }
    },
    "/api/asignacion-transporte/conductores/{id}": {
      "get": {
        "description": "Obtiene la información detallada de un conductor específico",
        "operationId": "AsignacionTransporteController_findConductorById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del conductor",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conductor encontrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConductorResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conductor no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener un conductor por ID",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "patch": {
        "description": "Actualiza la información de un conductor existente",
        "operationId": "AsignacionTransporteController_updateConductor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del conductor",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Datos del conductor a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConductorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Conductor actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ConductorResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos o conductor ya existe"
          },
          "404": {
            "description": "Conductor no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Actualizar un conductor",
        "tags": [
          "Asignación de Transporte"
        ]
      },
      "delete": {
        "description": "Elimina un conductor del sistema (cambia su estado a inactivo)",
        "operationId": "AsignacionTransporteController_removeConductor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID único del conductor",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Conductor eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EmptyResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Conductor no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Eliminar un conductor",
        "tags": [
          "Asignación de Transporte"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Transporte",
    "description": "API para el sistema de monitoreo de transporte",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingresa tu token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateVehiculoDto": {
        "type": "object",
        "properties": {
          "placa": {
            "type": "string",
            "description": "Placa del vehículo",
            "example": "ABC1234",
            "maxLength": 7
          },
          "marca": {
            "type": "string",
            "description": "Marca del vehículo",
            "example": "Toyota",
            "maxLength": 15
          },
          "modelo": {
            "type": "string",
            "description": "Modelo del vehículo",
            "example": "Hilux",
            "maxLength": 15
          },
          "año": {
            "type": "number",
            "description": "Año del vehículo",
            "example": 2020,
            "minimum": 2000
          },
          "color": {
            "type": "string",
            "description": "Color del vehículo",
            "example": "Blanco",
            "maxLength": 15
          },
          "tipo_vehiculo": {
            "type": "string",
            "description": "Tipo de vehículo",
            "enum": [
              "camion",
              "camioneta",
              "furgoneta"
            ],
            "example": "camioneta"
          },
          "capacidad_carga": {
            "type": "number",
            "description": "Capacidad de carga en toneladas",
            "example": 2.5,
            "minimum": 0
          },
          "capacidad_combustible": {
            "type": "number",
            "description": "Capacidad de combustible en litros",
            "example": 80,
            "minimum": 0
          },
          "estado_vehiculo": {
            "type": "boolean",
            "description": "Estado del vehículo",
            "example": true,
            "default": true
          },
          "numero_soat": {
            "type": "string",
            "description": "Número de SOAT",
            "example": "1234567",
            "maxLength": 7
          },
          "fecha_vencimiento_soat": {
            "type": "string",
            "description": "Fecha de vencimiento del SOAT",
            "example": "2024-12-31"
          },
          "gps_imei": {
            "type": "string",
            "description": "IMEI del GPS",
            "example": "123456789012345",
            "maxLength": 30
          },
          "gps_estado": {
            "type": "boolean",
            "description": "Estado del GPS",
            "example": true,
            "default": true
          },
          "foto_vehiculo": {
            "type": "string",
            "description": "URL de la foto del vehículo",
            "example": "https://example.com/foto.jpg"
          },
          "id_usuario": {
            "type": "string",
            "description": "ID del usuario que registra el vehículo",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "placa",
          "marca",
          "modelo",
          "año",
          "color",
          "tipo_vehiculo",
          "capacidad_carga",
          "capacidad_combustible",
          "estado_vehiculo",
          "numero_soat",
          "fecha_vencimiento_soat",
          "gps_imei",
          "gps_estado",
          "id_usuario"
        ]
      },
      "VehiculoListResponseDto": {
        "type": "object",
        "properties": {
          "vehiculos": {
            "description": "Lista de vehículos",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehiculoResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total de vehículos",
            "example": 10
          },
          "pagina": {
            "type": "number",
            "description": "Página actual",
            "example": 1
          },
          "por_pagina": {
            "type": "number",
            "description": "Elementos por página",
            "example": 10
          }
        },
        "required": [
          "vehiculos",
          "total",
          "pagina",
          "por_pagina"
        ]
      },
...
      "EmptyResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Estado de la respuesta",
            "example": true,
            "nullable": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de la respuesta",
            "example": "Operación exitosa",
            "nullable": false
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Estado de la respuesta",
            "example": false,
            "nullable": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error",
            "example": "Error interno del servidor",
            "nullable": false
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
    }
  }
}